+++ tests/SigstoreVerifierTest.php
@@ -11,6 +11,7 @@ use K2gl\Sigstore\CertificateAuthority;
 use K2gl\Sigstore\Exception\UnsupportedBundleException;
 use K2gl\Sigstore\Exception\VerificationFailedException;
 use K2gl\Sigstore\IdentityPolicy;
+use K2gl\Sigstore\Internal\Asn1;
 use K2gl\Sigstore\Internal\Certificate;
 use K2gl\Sigstore\Internal\CertificateChainVerifier;
 use K2gl\Sigstore\Internal\CertificateKeyVerifier;
@@ -18,6 +19,8 @@ use K2gl\Sigstore\Internal\Ecdsa;
 use K2gl\Sigstore\Internal\Json;
 use K2gl\Sigstore\Internal\Pem;
 use K2gl\Sigstore\Internal\RekorVerifier;
+use K2gl\Sigstore\Internal\Sct;
+use K2gl\Sigstore\Internal\SctVerifier;
 use K2gl\Sigstore\Internal\TrustRootJson;
 use K2gl\Sigstore\SigstoreVerifier;
 use K2gl\Sigstore\TlogEntry;
@@ -43,6 +46,9 @@ use PHPUnit\Framework\TestCase;
 #[CoversClass(CertificateChainVerifier::class)]
 #[CoversClass(CertificateKeyVerifier::class)]
 #[CoversClass(RekorVerifier::class)]
+#[CoversClass(SctVerifier::class)]
+#[CoversClass(Sct::class)]
+#[CoversClass(Asn1::class)]
 #[CoversClass(Ecdsa::class)]
 #[CoversClass(Json::class)]
 #[CoversClass(TrustRootJson::class)]